Move common files to utils
