+++ frontend_new/app/src/app/utils/validators.tsx
@@ -1,15 +1,20 @@
+
+/** Returns true on failed validation. Checks that string is not empty*/
 export const validateString = (value: string): boolean => {
   return value.length === 0;
 }
 
+/** Returns true on failed validation*/
 export const validatePassword = (value: string): boolean => {
   return value.length < 6;
 }
 
+/** Returns true on failed validation*/
 export const validateEmail = (value: string): boolean => {
   return !/^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/.test(value);
 }
 
+/** Returns true on failed validation*/
 export const validateConfirmPassword = (password: string, confirmPassword: string): boolean => {
   return password !== confirmPassword;
 }
@@ -19,6 +24,7 @@ interface validateNumberStringOptions {
   decimalDigits?: number;
 }
 
+/** Returns true on failed validation*/
 export const validateNumberString = (value: string, options: validateNumberStringOptions): boolean => {
   let res = true;
   const valueToNumber = Number(value);
